Implement swagger 2.0 multipart formdata parsing
There are several ways to do this, we use the simplest:
Parse some chunk of data into memory, then parse all parts
separately and process them. This has the benefit that we can
retain the structure of the parsing quite easily and shift most
of the work to the implementation.

Limitations:
- The request parses a maximum of 32 MB. This helps against DDoS
  attacks on memory, so it's not too bad
- The minimum memory footprint may be higher than necessary